+++ spec/models/activity_spec.rb
@@ -92,6 +92,40 @@ RSpec.describe Activity, type: :model do
       end
     end
 
+    context "when the actual_start_date is not blank" do
+      it "allows todays date" do
+        activity = build(:activity, actual_start_date: Date.today)
+        expect(activity.valid?).to be_truthy
+      end
+
+      it "allows dates in the past" do
+        activity = build(:activity, actual_start_date: 1.year.ago)
+        expect(activity.valid?).to be_truthy
+      end
+
+      it "does not allow a date in the future" do
+        activity = build(:activity, actual_start_date: 1.day.from_now)
+        expect(activity.valid?).to be_falsey
+      end
+    end
+
+    context "when the actual_end_date is not blank" do
+      it "allows todays date" do
+        activity = build(:activity, actual_end_date: Date.today)
+        expect(activity.valid?).to be_truthy
+      end
+
+      it "allows dates in the past" do
+        activity = build(:activity, actual_end_date: 1.year.ago)
+        expect(activity.valid?).to be_truthy
+      end
+
+      it "does not allow a date in the future" do
+        activity = build(:activity, actual_end_date: 1.day.from_now)
+        expect(activity.valid?).to be_falsey
+      end
+    end
+
     context "when recipient_region is blank" do
       subject { build(:activity, recipient_region: nil, wizard_status: :country) }
       it { should validate_presence_of(:recipient_region) }