+++ CHANGELOG.md
@@ -54,3 +54,4 @@
 - Users can edit a reported implementing organisation
 - Users can view implementing organisations in the Activity XML
 - Planned start and end dates are mandatory
+- Actual start and end dates must not be in the future